# Chore: use native `Object.assign` (refs #6407)