--- a/lib/cli-engine.js
+++ b/lib/cli-engine.js
@@ -17,9 +17,6 @@
 
 const fs = require("fs"),
     path = require("path"),
-
-    lodash = require("lodash"),
-
     rules = require("./rules"),
     eslint = require("./eslint"),
     defaultOptions = require("../conf/cli-options"),
@@ -436,7 +433,7 @@ function getCacheFile(cacheFile, cwd) {
  */
 function CLIEngine(options) {
 
-    options = lodash.assign(
+    options = Object.assign(
         Object.create(null),
         defaultOptions,
         {cwd: process.cwd()},