--- a/lib/config/autoconfig.js
+++ b/lib/config/autoconfig.js
@@ -172,7 +172,7 @@ Registry.prototype = {
         const ruleIds = Object.keys(this.rules),
             newRegistry = new Registry();
 
-        newRegistry.rules = lodash.assign({}, this.rules);
+        newRegistry.rules = Object.assign({}, this.rules);
         ruleIds.forEach(function(ruleId) {
             const errorFreeItems = newRegistry.rules[ruleId].filter(function(registryItem) {
                 return (registryItem.errorCount === 0);
@@ -197,7 +197,7 @@ Registry.prototype = {
         const ruleIds = Object.keys(this.rules),
             newRegistry = new Registry();
 
-        newRegistry.rules = lodash.assign({}, this.rules);
+        newRegistry.rules = Object.assign({}, this.rules);
         ruleIds.forEach(function(ruleId) {
             newRegistry.rules[ruleId] = newRegistry.rules[ruleId].filter(function(registryItem) {
                 return (typeof registryItem.errorCount !== "undefined");
@@ -260,7 +260,7 @@ Registry.prototype = {
         const ruleIds = Object.keys(this.rules),
             newRegistry = new Registry();
 
-        newRegistry.rules = lodash.assign({}, this.rules);
+        newRegistry.rules = Object.assign({}, this.rules);
         ruleIds.forEach(function(ruleId) {
             newRegistry.rules[ruleId] = this.rules[ruleId].filter(function(registryItem) {
                 return (registryItem.specificity === specificity);
@@ -283,7 +283,7 @@ Registry.prototype = {
             lintedRegistry;
 
         lintedRegistry = new Registry();
-        lintedRegistry.rules = lodash.assign({}, this.rules);
+        lintedRegistry.rules = Object.assign({}, this.rules);
 
         const ruleSets = lintedRegistry.buildRuleSets();
 
@@ -300,7 +300,7 @@ Registry.prototype = {
             ruleSetIdx = 0;
 
             ruleSets.forEach(function(ruleSet) {
-                const lintConfig = lodash.assign({}, config, {rules: ruleSet});
+                const lintConfig = Object.assign({}, config, {rules: ruleSet});
                 const lintResults = eslint.verify(sourceCodes[filename], lintConfig);
 
                 lintResults.forEach(function(result) {
@@ -338,7 +338,7 @@ Registry.prototype = {
  * @returns {Object}        config object using `"extends": "eslint:recommended"`
  */
 function extendFromRecommended(config) {
-    const newConfig = lodash.assign({}, config);
+    const newConfig = Object.assign({}, config);
 
     ConfigOps.normalizeToStrings(newConfig);