--- a/lib/config/config-initializer.js
+++ b/lib/config/config-initializer.js
@@ -10,7 +10,6 @@
 //------------------------------------------------------------------------------
 
 const util = require("util"),
-    lodash = require("lodash"),
     inquirer = require("inquirer"),
     ProgressBar = require("progress"),
     autoconfig = require("./autoconfig.js"),
@@ -112,7 +111,7 @@ function installModules(config) {
 function configureRules(answers, config) {
     const BAR_TOTAL = 20,
         BAR_SOURCE_CODE_TOTAL = 4,
-        newConfig = lodash.assign({}, config),
+        newConfig = Object.assign({}, config),
         disabledConfigs = {};
     let sourceCodes,
         registry;
@@ -187,7 +186,7 @@ function configureRules(answers, config) {
     const defaultConfigs = registry.filterBySpecificity(1).createConfig().rules;
 
     // Combine configs in reverse priority order (later take precedence)
-    newConfig.rules = lodash.assign({}, disabledConfigs, defaultConfigs, specThreeConfigs, specTwoConfigs, singleConfigs);
+    newConfig.rules = Object.assign({}, disabledConfigs, defaultConfigs, specThreeConfigs, specTwoConfigs, singleConfigs);
 
     // Make sure progress bar has finished (floating point rounding)
     bar.update(BAR_TOTAL);
@@ -410,7 +409,7 @@ function promptUser(callback) {
             // early exit if you are using automatic style generation
             if (earlyAnswers.source === "auto") {
                 try {
-                    const combinedAnswers = lodash.assign({}, earlyAnswers, secondAnswers);
+                    const combinedAnswers = Object.assign({}, earlyAnswers, secondAnswers);
 
                     config = processAnswers(combinedAnswers);
                     installModules(config);
@@ -460,7 +459,7 @@ function promptUser(callback) {
                 }
             ], function(answers) {
                 try {
-                    const totalAnswers = lodash.assign({}, earlyAnswers, secondAnswers, answers);
+                    const totalAnswers = Object.assign({}, earlyAnswers, secondAnswers, answers);
 
                     config = processAnswers(totalAnswers);
                     installModules(config);