--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -13,7 +13,6 @@ const assert = require("assert"),
     EventEmitter = require("events").EventEmitter,
     escope = require("escope"),
     levn = require("levn"),
-    lodash = require("lodash"),
     blankScriptAST = require("../conf/blank-script.json"),
     DEFAULT_PARSER = require("../conf/eslint.json").parser,
     replacements = require("../conf/replacements.json"),
@@ -152,7 +151,7 @@ function addDeclaredGlobals(program, globalScope, config) {
         explicitGlobals = {},
         builtin = Environments.get("builtin");
 
-    lodash.assign(declaredGlobals, builtin);
+    Object.assign(declaredGlobals, builtin);
 
     Object.keys(config.env).forEach(function(name) {
         if (config.env[name]) {
@@ -160,14 +159,14 @@ function addDeclaredGlobals(program, globalScope, config) {
                 environmentGlobals = env && env.globals;
 
             if (environmentGlobals) {
-                lodash.assign(declaredGlobals, environmentGlobals);
+                Object.assign(declaredGlobals, environmentGlobals);
             }
         }
     });
 
-    lodash.assign(exportedGlobals, config.exported);
-    lodash.assign(declaredGlobals, config.globals);
-    lodash.assign(explicitGlobals, config.astGlobals);
+    Object.assign(exportedGlobals, config.exported);
+    Object.assign(declaredGlobals, config.globals);
+    Object.assign(explicitGlobals, config.astGlobals);
 
     Object.keys(declaredGlobals).forEach(function(name) {
         let variable = globalScope.set.get(name);
@@ -325,16 +324,16 @@ function modifyConfigsFromComments(filename, ast, config, reportingConfig, messa
             if (comment.type === "Block") {
                 switch (match[1]) {
                     case "exported":
-                        lodash.assign(commentConfig.exported, parseBooleanConfig(value, comment));
+                        Object.assign(commentConfig.exported, parseBooleanConfig(value, comment));
                         break;
 
                     case "globals":
                     case "global":
-                        lodash.assign(commentConfig.astGlobals, parseBooleanConfig(value, comment));
+                        Object.assign(commentConfig.astGlobals, parseBooleanConfig(value, comment));
                         break;
 
                     case "eslint-env":
-                        lodash.assign(commentConfig.env, parseListConfig(value));
+                        Object.assign(commentConfig.env, parseListConfig(value));
                         break;
 
                     case "eslint-disable":
@@ -379,7 +378,7 @@ function modifyConfigsFromComments(filename, ast, config, reportingConfig, messa
             commentConfig = ConfigOps.merge(commentConfig, env);
         }
     });
-    lodash.assign(commentConfig.rules, commentRules);
+    Object.assign(commentConfig.rules, commentRules);
 
     return ConfigOps.merge(config, commentConfig);
 }
@@ -527,7 +526,7 @@ function findEslintEnv(text) {
     eslintEnvPattern.lastIndex = 0;
 
     while ((match = eslintEnvPattern.exec(text))) {
-        retv = lodash.assign(retv || {}, parseListConfig(match[1]));
+        retv = Object.assign(retv || {}, parseListConfig(match[1]));
     }
 
     return retv;
@@ -612,7 +611,7 @@ module.exports = (function() {
 
         // merge in any additional parser options
         if (config.parserOptions) {
-            parserOptions = lodash.assign({}, config.parserOptions, parserOptions);
+            parserOptions = Object.assign({}, config.parserOptions, parserOptions);
         }
 
         /*
@@ -740,10 +739,10 @@ module.exports = (function() {
 
         if (envInFile) {
             if (!config || !config.env) {
-                config = lodash.assign({}, config || {}, {env: envInFile});
+                config = Object.assign({}, config || {}, {env: envInFile});
             } else {
-                config = lodash.assign({}, config);
-                config.env = lodash.assign({}, config.env, envInFile);
+                config = Object.assign({}, config);
+                config.env = Object.assign({}, config.env, envInFile);
             }
         }