--- a/lib/ignored-paths.js
+++ b/lib/ignored-paths.js
@@ -9,8 +9,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const lodash = require("lodash"),
-    fs = require("fs"),
+const fs = require("fs"),
     path = require("path"),
     ignore = require("ignore"),
     pathUtil = require("./util/path-util");
@@ -58,7 +57,7 @@ function findIgnoreFile(cwd) {
  */
 function mergeDefaultOptions(options) {
     options = (options || {});
-    return lodash.assign({}, DEFAULT_OPTIONS, options);
+    return Object.assign({}, DEFAULT_OPTIONS, options);
 }
 
 //------------------------------------------------------------------------------