--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -12,7 +12,6 @@
 // Rule Definition
 //------------------------------------------------------------------------------
 const util = require("util");
-const lodash = require("lodash");
 
 module.exports = {
     meta: {
@@ -123,7 +122,7 @@ module.exports = {
                         const: variableDeclaratorRules
                     };
                 } else if (typeof variableDeclaratorRules === "object") {
-                    lodash.assign(options.VariableDeclarator, variableDeclaratorRules);
+                    Object.assign(options.VariableDeclarator, variableDeclaratorRules);
                 }
 
                 if (typeof opts.outerIIFEBody === "number") {