--- a/lib/rules/new-cap.js
+++ b/lib/rules/new-cap.js
@@ -9,8 +9,6 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const lodash = require("lodash");
-
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
@@ -115,7 +113,7 @@ module.exports = {
 
     create: function(context) {
 
-        const config = context.options[0] ? lodash.assign({}, context.options[0]) : {};
+        const config = context.options[0] ? Object.assign({}, context.options[0]) : {};
 
         config.newIsCap = config.newIsCap !== false;
         config.capIsNew = config.capIsNew !== false;