--- a/lib/rules/operator-linebreak.js
+++ b/lib/rules/operator-linebreak.js
@@ -5,8 +5,7 @@
 
 "use strict";
 
-const lodash = require("lodash"),
-    astUtils = require("../ast-utils");
+const astUtils = require("../ast-utils");
 
 //------------------------------------------------------------------------------
 // Rule Definition
@@ -47,7 +46,7 @@ module.exports = {
         const usedDefaultGlobal = !context.options[0];
         const globalStyle = context.options[0] || "after";
         const options = context.options[1] || {};
-        const styleOverrides = options.overrides ? lodash.assign({}, options.overrides) : {};
+        const styleOverrides = options.overrides ? Object.assign({}, options.overrides) : {};
 
         if (usedDefaultGlobal && !styleOverrides["?"]) {
             styleOverrides["?"] = "before";