--- a/lib/rules/require-jsdoc.js
+++ b/lib/rules/require-jsdoc.js
@@ -4,8 +4,6 @@
  */
 "use strict";
 
-const lodash = require("lodash");
-
 module.exports = {
     meta: {
         docs: {
@@ -46,7 +44,7 @@ module.exports = {
             MethodDefinition: false,
             ClassDeclaration: false
         };
-        const options = lodash.assign(DEFAULT_OPTIONS, context.options[0] && context.options[0].require || {});
+        const options = Object.assign(DEFAULT_OPTIONS, context.options[0] && context.options[0].require || {});
 
         /**
          * Report the error message