--- a/lib/rules/strict.js
+++ b/lib/rules/strict.js
@@ -9,8 +9,6 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const lodash = require("lodash");
-
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
@@ -231,7 +229,7 @@ module.exports = {
         };
 
         if (mode === "function") {
-            lodash.assign(rule, {
+            Object.assign(rule, {
 
                 // Inside of class bodies are always strict mode.
                 ClassBody: function() {