--- a/lib/util/module-resolver.js
+++ b/lib/util/module-resolver.js
@@ -9,8 +9,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const lodash = require("lodash"),
-    Module = require("module");
+const Module = require("module");
 
 //------------------------------------------------------------------------------
 // Private
@@ -39,7 +38,7 @@ const DEFAULT_OPTIONS = {
 function ModuleResolver(options) {
     options = options || {};
 
-    this.options = lodash.assign({}, DEFAULT_OPTIONS, options);
+    this.options = Object.assign({}, DEFAULT_OPTIONS, options);
 }
 
 ModuleResolver.prototype = {