--- a/lib/util/source-code-util.js
+++ b/lib/util/source-code-util.js
@@ -9,8 +9,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const lodash = require("lodash"),
-    CLIEngine = require("../cli-engine"),
+const CLIEngine = require("../cli-engine"),
     eslint = require("../eslint"),
     globUtil = require("./glob-util"),
     baseDefaultOptions = require("../../conf/cli-options");
@@ -30,7 +29,7 @@ const debug = require("debug")("eslint:source-code-util");
  */
 function getSourceCodeOfFile(filename, options) {
     debug("getting sourceCode of", filename);
-    const opts = lodash.assign({}, options, { rules: {}});
+    const opts = Object.assign({}, options, { rules: {}});
     const cli = new CLIEngine(opts);
     const results = cli.executeOnFiles([filename]);
 
@@ -72,7 +71,7 @@ function getSourceCodeOfFiles(patterns, options, cb) {
         patterns = [patterns];
     }
 
-    const defaultOptions = lodash.assign({}, baseDefaultOptions, {cwd: process.cwd()});
+    const defaultOptions = Object.assign({}, baseDefaultOptions, {cwd: process.cwd()});
 
     if (typeof options === "undefined") {
         opts = defaultOptions;
@@ -80,7 +79,7 @@ function getSourceCodeOfFiles(patterns, options, cb) {
         cb = options;
         opts = defaultOptions;
     } else if (typeof options === "object") {
-        opts = lodash.assign({}, defaultOptions, options);
+        opts = Object.assign({}, defaultOptions, options);
     }
     debug("constructed options:", opts);
     patterns = globUtil.resolveFileGlobPatterns(patterns, opts);