--- a/lib/util/source-code.js
+++ b/lib/util/source-code.js
@@ -8,8 +8,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const lodash = require("lodash"),
-    createTokenStore = require("../token-store.js"),
+const createTokenStore = require("../token-store.js"),
     Traverser = require("./traverser");
 
 //------------------------------------------------------------------------------
@@ -281,7 +280,7 @@ SourceCode.prototype = {
             }
         });
 
-        return result ? lodash.assign({parent: resultParent}, result) : null;
+        return result ? Object.assign({parent: resultParent}, result) : null;
     },
 
     /**