<!--
Thanks for submitting a pull request to ESLint. Before continuing, please be sure you've read over our guidelines:
http://eslint.org/docs/developer-guide/contributing/pull-requests

Specifically, all pull requests containing code require an **accepted** issue (documentation-only pull requests do not require an issue). If this pull request contains code and there isn't yet an issue explaining why you're submitting this pull request, please stop and open a new issue first.

Please answer all questions below.
-->

**What issue does this pull request address?**
Use native function rather than a polyfill from `lodash`. This way we are more native bound than library bound :smile: 

**What changes did you make? (Give an overview)**
Replaced `lodash.assign` with native `Object.assign`. As a result of this i was able to get rid of `lodash` in many modules.

**Is there anything you'd like reviewers to focus on?**

nothing specific