@ilyavolodin 2016-08-02
@gyandeeps Can you run perf test before and after for this change please?