@gyandeeps 2016-08-02
@ilyavolodin 

### Before
```sh
Loading:
  Load performance Run #1:  285.152035ms
  Load performance Run #2:  293.246393ms
  Load performance Run #3:  286.062485ms
  Load performance Run #4:  288.535803ms
  Load performance Run #5:  277.653648ms

  Load Performance median:  286.062485ms


Single File:
  CPU Speed is 2793 with multiplier 13000000
  Performance Run #1:  3404.311078ms
  Performance Run #2:  3374.370675ms
  Performance Run #3:  3362.172693ms
  Performance Run #4:  3387.370983ms
  Performance Run #5:  3407.566561ms

  Performance budget ok:  3387.370983ms (limit: 4654.493376297887ms)


Multi Files (0 files):
  CPU Speed is 2793 with multiplier 39000000
  Performance Run #1:  944.787268ms
  Performance Run #2:  1000.370191ms
  Performance Run #3:  1036.133591ms
  Performance Run #4:  1156.578407ms
  Performance Run #5:  1120.515556ms

  Performance budget ok:  1036.133591ms (limit: 13963.480128893663ms)
```

### After
```sh
Loading:
  Load performance Run #1:  283.44549ms
  Load performance Run #2:  281.234239ms
  Load performance Run #3:  285.52076ms
  Load performance Run #4:  290.018767ms
  Load performance Run #5:  279.19379ms

  Load Performance median:  283.44549ms


Single File:
  CPU Speed is 2793 with multiplier 13000000
  Performance Run #1:  3345.934495ms
  Performance Run #2:  3364.047106ms
  Performance Run #3:  3374.118139ms
  Performance Run #4:  3369.067411ms
  Performance Run #5:  3377.237274ms

  Performance budget ok:  3369.067411ms (limit: 4654.493376297887ms)


Multi Files (0 files):
  CPU Speed is 2793 with multiplier 39000000
  Performance Run #1:  901.099578ms
  Performance Run #2:  900.602568ms
  Performance Run #3:  901.339652ms
  Performance Run #4:  894.463259ms
  Performance Run #5:  893.263986ms

  Performance budget ok:  900.602568ms (limit: 13963.480128893663ms)
```